requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
apturl==0.5.2
asttokens==3.0.0
astunparse==1.6.3
Authlib==1.3.2
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
cachelib==0.13.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.4.0
chex==0.1.87
cliapp==1.20180812.1
click==7.1.2
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cssselect==1.2.0
cssutils==2.11.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
etils==1.11.0
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.14.1
filelock==3.16.1
Flask==1.1.2
Flask-Cors==5.0.0
Flask-OAuthlib==0.9.6
flask-talisman==1.1.0
flatbuffers==24.3.25
flax==0.10.2
fsspec==2024.10.0
future==0.18.2
gast==0.6.0
google-auth==1.27.0
gunicorn==20.1.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.0
huggingface-hub==0.26.3
humanize==4.11.0
icecream==2.1.3
idna==2.10
importlib-metadata==4.6.4
importlib_resources==6.4.5
itsdangerous==1.1.0
jax==0.4.36
jaxlib==0.4.36
jeepney==0.7.1
Jinja2==2.11.3
jiter==0.8.0
joblib==1.4.2
keras==3.7.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==1.1.1
mdurl==0.1.2
ml-dtypes==0.4.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.1.0
olefile==0.46
openai==1.57.0
opt_einsum==3.4.0
optax==0.2.4
optree==0.13.1
orbax-checkpoint==0.10.2
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pinecone==5.4.1
pinecone-client==5.0.1
pinecone-plugin-assistant==0.4.0
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
premailer==3.10.0
protobuf==5.29.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2024.11.6
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.0
rich==13.9.4
rsa==4.7.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.1
sendgrid==6.11.0
sentence-transformers==3.3.1
simplejson==3.19.3
six==1.15.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starkbank-ecdsa==2.2.0
sympy==1.13.1
systemd-python==234
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
tensorstore==0.1.69
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.0
triton==3.1.0
ttystatus==0.38
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.3
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==1.0.1
wrapt==1.17.0
xdg==5
xkit==0.0.0
yagmail==0.15.293
zipp==1.0.0
